Moved DirectiveLocations to ComponentModel
